# swagger.yaml
swagger: '2.0'
info:
  title: App-container API
  description: >-
    App-container APIs mainly consists of authentication, resource transaction
    and order transaction.
  version: 1.0.1
  termsOfService: 'http://console.hnabaas.com/terms'
  contact:
    name: API Support
    email: ct.shi@hnair.com
host: hnabaas.com
basePath: /service
schemes:
  - http
  - https
produces:
  - application/json
paths:
  /resource/vehicle:
    post:
      tags:
        - Vehicle
      summary: add a new vehicle to the ledger
      description: ''
      operationId: addVehicle
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: Vehicle object that needs to be added to the ledger
          required: true
          schema:
            $ref: '#/definitions/Vehicle'
      responses:
        '405':
          description: Invalid input
    put:
      tags:
        - Vehicle
      summary: Update an existing vehicle
      description: ''
      operationId: updateVehicle
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - in: body
          name: body
          description: Vehicle object that needs to be added to the ledger
          required: true
          schema:
            $ref: '#/definitions/Vehicle'
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Vehicle not found
        '405':
          description: Validation exception
  '/resource/vehicle/{vehicleId}':
    get:
      tags:
        - Vehicle
      summary: Find vehicle by id
      description: Returns a single vehicle
      operationId: getVehicleById
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passsed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: vehicleId
          in: path
          description: ID of a vehicle to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Vehicle'
        '400':
          description: Invalid ID supplied
        '404':
          description: Vehicle not found
    delete:
      tags:
        - Vehicle
      summary: Delete a vehicle by id
      description: ''
      operationId: deleteVehicle
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: vehicleId
          in: path
          description: ID of a vehicle to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Vehicle'
        '400':
          description: Invalid ID supplied
        '404':
          description: Vehicle not found
  /resource/vehicle/findByOwnerId:
    get:
      tags:
        - Vehicle
      summary: Finds vehicles by ownerId
      description: ''
      operationId: findVehiclesByOwnerId
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: ownerid
          in: query
          description: OwnerId values that need to be considered for filter
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Vehicle'
        '400':
          description: Invalid ownerId value
...
  /resource/container:
    post:
      tags:
        - Container
      summary: add a new container to the ledger
      description: ''
      operationId: addContainer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: Container object that needs to be added to the ledger
          required: true
          schema:
            $ref: '#/definitions/Container'
      responses:
        '405':
          description: Invalid input
    put:
      tags:
        - Container
      summary: Update an existing container
      description: ''
      operationId: updateContainer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - in: body
          name: body
          description: Container object that needs to be added to the ledger
          required: true
          schema:
            $ref: '#/definitions/Container'
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Container not found
        '405':
          description: Validation exception
  '/resource/container/{containerId}':
    get:
      tags:
        - Container
      summary: Find container by id
      description: Returns a single container
      operationId: getContainerById
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: containerId
          in: path
          description: ID of a shipping schedule to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Container'
        '400':
          description: Invalid ID supplied
        '404':
          description: Container not found
    delete:
      tags:
        - Container
      summary: Delete a container by id
      description: ''
      operationId: deleteContainer
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: containerId
          in: path
          description: ID of a shippingSchedule to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/ShippingSchedule'
        '400':
          description: Invalid ID supplied
        '404':
          description: Vehicle not found
  /resource/container/findByOwnerId:
    get:
      tags:
        - Container
      summary: Finds containers by ownerId
      description: ''
      operationId: findContainersByOwnerId
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: ownerid
          in: query
          description: OwnerId values that need to be considered for filter
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Container'
        '400':
          description: Invalid ownerId value
  '/resource/transporttask/{taskId}':
    get:
      tags:
        - TransportTask
      summary: Find transport task by id
      description: Returns a single transport task
      operationId: getTransportTaskById
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: taskId
          in: path
          description: ID of a transport task to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/TransportTask'
        '400':
          description: Invalid ID supplied
        '404':
          description: Container not found
    delete:
      tags:
        - TransportTask
      summary: Delete a transport task by id
      description: ''
      operationId: deleteTransportTask
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: taskId
          in: path
          description: ID of a transport task
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/TransportTask'
        '400':
          description: Invalid ID supplied
        '404':
          description: TransportTask not found
  /resource/transporttask/findByOwnerId:
    get:
      tags:
        - TransportTask
      summary: Finds TransportTasks by ownerId
      description: ''
      operationId: findTransportTasksByOwnerId
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: ownerid
          in: query
          description: OwnerId values that need to be considered for filter
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TransportTask'
        '400':
          description: Invalid ownerId value
  /order/client/create:
    post:
      tags:
        - Order
      summary: 'add a new order to the ledger, proposed by the client'
      description: ''
      operationId: createOrder
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: OrderCreateRequest object that needs to be added to the ledger
          required: true
          schema:
            $ref: '#/definitions/OrderCreateRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/cargoagent/check:
    post:
      tags:
        - Order
      summary: 'check to accept the order or not, proposed by the cargo agent'
      description: ''
      operationId: checkOrder
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderCheckRequest object that needs to be used to accept or deny
            order
          required: true
          schema:
            $ref: '#/definitions/OrderCheckRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/cargoagent/bookspace:
    post:
      tags:
        - Order
      summary: 'book space from a shipper, proposed by the cargo agent'
      description: ''
      operationId: bookSpace
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: OrderBookSpaceRequest object that needs to be used to book space
          required: true
          schema:
            $ref: '#/definitions/OrderBookSpaceRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/cargoagent/bookvehicle:
    post:
      tags:
        - Order
      summary: 'book vehicle from a carrier, proposed by the cargo agent'
      description: ''
      operationId: bookVehicle
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: OrderBookVehicleRequest object that needs to be used to book space
          required: true
          schema:
            $ref: '#/definitions/OrderBookVehicleRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/carrier/fetchemptycontainers:
    post:
      tags:
        - Order
      summary: 'fetch empty containers from a shipper, proposed by the carrier'
      description: ''
      operationId: fetchEmptyContainers
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderFetchEmptyContainersRequest object that used to add the date
            for delivery， 记录设备（空集装箱）移出时间
          required: true
          schema:
            $ref: '#/definitions/OrderFetchEmptyContainersRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/carrier/packgoods:
    post:
      tags:
        - Order
      summary: 'pack goods into the empty containers, proposed by the carrier'
      description: ''
      operationId: packGoods
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderPackGoodsRequest object that used to add the the packing list
            to the order
          required: true
          schema:
            $ref: '#/definitions/OrderPackGoodsRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/carrier/arriveyard:
    post:
      tags:
        - Order
      summary: 'deliver the goods and containers to yard , proposed by the carrier'
      description: ''
      operationId: arriveYard
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderArriveYardRequest object that used to add the the date for
            arriving the yard. 抵达堆场后该运输任务结束
          required: true
          schema:
            $ref: '#/definitions/OrderArriveYardRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/shipper/loadgoods:
    post:
      tags:
        - Order
      summary: Goods being loaded into the ship
      description: ''
      operationId: loadGoods
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderLoadGoods object that used to add the the date for loading
            ship. 货物装船时间
          required: true
          schema:
            $ref: '#/definitions/OrderLoadGoodsRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/shipper/departrue:
    post:
      tags:
        - Order
      summary: The ship departed from the loading port.
      description: ''
      operationId: departure
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderDepatureRequest object that used to add the the date for
            departure.集装箱离港时间
          required: true
          schema:
            $ref: '#/definitions/OrderDepartureRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
  /order/shipper/arrivedestinationport:
    post:
      tags:
        - Order
      summary: The ship arrived the destination port.
      description: ''
      operationId: arriveDestinationPort
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderArriveDestinationRequest object that used to add the the date
            for its arrival.集装箱到目的港口时间
          required: true
          schema:
            $ref: '#/definitions/OrderArriveDestinationPortRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
...
  /order/client/confirmreceipt:
    post:
      tags:
        - Order
      summary: The clent confirmed that the goods has been receipt.
      description: ''
      operationId: confirmReceipt
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: header
          name: userid
          description: user id to be passed as a header
          required: true
          type: string
        - in: header
          name: token
          description: token to be passed as a header
          required: true
          type: string
        - in: header
          name: sessionid
          description: session id to be passed as a header
          required: true
          type: string
        - name: body
          in: body
          description: >-
            OrderConfirmReceiptRequest object that used to add the the date for
            receipt.客户确认收到货物
          required: true
          schema:
            $ref: '#/definitions/OrderConfirmReceiptRequest'
      responses:
        '200':
          description: Successful Operation
        '405':
          description: Invalid input
...
  /message:
    post:
      tags:
        - Message
      summary: add a new message to the ledger
      description: ''
      operationId: addMessage
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Message object that needs to be added to the mysql
          required: true
          schema:
            $ref: '#/definitions/Message'
      responses:
        '405':
          description: Invalid input
...
definitions:
  Vehicle:
    type: object
    properties:
      Id:
        type: string
        description: 车辆的id，系统生成
      VehicleNo:
        type: string
        description: 车辆编号
      Driver:
        $ref: '#/definitions/Driver'
      Status:
        type: string
        description: 车辆的当前使用状态，空闲或者
      OwnerId:
        type: string
        description: 所属拖车公司的ID
...
  ShippingSchedule:
    type: object
    properties:
      Id:
        type: string
        description: 航次的id，系统生成
      OwnerId:
        type: string
        description: 所属船公司Id
      VoyNo:
        type: string
        description: 航次号
      Vessel:
        $ref: '#/definitions/Vessel'
      PortOfLoading:
        type: string
        description: 装船港口
      PortOfDischarge:
        type: string
        description: 卸货港口
      PortOfDelivery:
        type: string
        description: ''
      DepartureDate:
        type: string
      ArrivalDate:
        type: string
      Status:
        type: string
        description: ''
...
  Container:
    type: object
    properties:
      Id:
        type: string
        description: 集装箱的id，系统生成
      ContainerNo:
        type: string
        description: 集装箱编号，系统生成
      Type:
        type: string
      MaxWeight:
        type: integer
        description: ''
      TareWeight:
        type: integer
        description: ''
      Measurement:
        type: integer
      Location:
        type: string
      Status:
        type: string
      OwnerId:
        type: string
...
  OrderCreateRequest:
    type: object
    properties:
      ConsigningForm:
        $ref: '#/definitions/ConsigningForm'
...